refactor(runtime): centralize scalar descriptor decoding
