+++ packages/runtime/src/mutation/postgres-program-codec.ts
@@ -1,30 +1,9 @@
 import { decodeRelationalScalar, type ScalarCodecV1 } from "../relational";
-import type { RecordValue } from "./postgres-program-types";
 
 function fail(message: string): never {
 	throw new TypeError(
 		`Invalid PostgreSQL Collection Operation plan: ${message}`,
 	);
-}
-
-function record(value: unknown, label: string): RecordValue {
-	if (!value || typeof value !== "object" || Array.isArray(value))
-		fail(`${label} must be an object`);
-	return value as RecordValue;
-}
-
-function exact(
-	value: RecordValue,
-	keys: readonly string[],
-	label: string,
-): void {
-	const actual = Object.keys(value).sort();
-	const expected = [...keys].sort();
-	if (
-		actual.length !== expected.length ||
-		expected.some((key, index) => key !== actual[index])
-	)
-		fail(`${label} has invalid keys`);
 }
 
 export function decodePostgresStatement(value: unknown, label: string): string {
@@ -39,133 +18,6 @@ export function decodePostgresStatement(value: unknown, label: string): string {
 	)
 		fail(`${label} is not one static statement`);
 	return value;
-}
-
-function nullableInteger(value: unknown, label: string): number | null {
-	if (value === null) return null;
-	if (!Number.isSafeInteger(value)) fail(`${label} is invalid`);
-	return value as number;
-}
-
-function boundedNullableInteger(
-	value: unknown,
-	label: string,
-	minimum: number,
-	maximum: number,
-): number | null {
-	const result = nullableInteger(value, label);
-	if (result !== null && (result < minimum || result > maximum))
-		fail(`${label} bounds are invalid`);
-	return result;
-}
-
-function nullableBigint(value: unknown, label: string): string | null {
-	if (value === null) return null;
-	if (
-		typeof value !== "string" ||
-		!/^(?:0|-[1-9][0-9]*|[1-9][0-9]*)$/.test(value)
-	)
-		fail(`${label} bigint is invalid`);
-	const result = BigInt(value);
-	if (
-		result < -9_223_372_036_854_775_808n ||
-		result > 9_223_372_036_854_775_807n
-	)
-		fail(`${label} bigint is invalid`);
-	return value;
-}
-
-export function decodePostgresScalarCodec(
-	value: unknown,
-	label: string,
-): ScalarCodecV1 {
-	const source = record(value, label);
-	if (
-		source.kind === "uuid" ||
-		source.kind === "boolean" ||
-		source.kind === "date"
-	) {
-		exact(source, ["kind"], label);
-		return Object.freeze({ kind: source.kind });
-	}
-	if (source.kind === "text") {
-		exact(source, ["kind", "minLength", "maxLength", "collation"], label);
-		const minLength = nullableInteger(source.minLength, `${label} minLength`);
-		const maxLength = nullableInteger(source.maxLength, `${label} maxLength`);
-		if (
-			(minLength !== null && minLength < 0) ||
-			(maxLength !== null && maxLength < 0) ||
-			(minLength !== null && maxLength !== null && minLength > maxLength) ||
-			source.collation !== "questpie.binary"
-		)
-			fail(`${label} bounds are invalid`);
-		return Object.freeze({
-			kind: "text",
-			minLength,
-			maxLength,
-			collation: "questpie.binary",
-		});
-	}
-	if (source.kind === "integer") {
-		exact(source, ["kind", "minimum", "maximum"], label);
-		const minimum = boundedNullableInteger(
-			source.minimum,
-			`${label} minimum`,
-			-2_147_483_648,
-			2_147_483_647,
-		);
-		const maximum = boundedNullableInteger(
-			source.maximum,
-			`${label} maximum`,
-			-2_147_483_648,
-			2_147_483_647,
-		);
-		if (minimum !== null && maximum !== null && minimum > maximum)
-			fail(`${label} bounds are invalid`);
-		return Object.freeze({ kind: "integer", minimum, maximum });
-	}
-	if (source.kind === "bigint") {
-		exact(source, ["kind", "minimum", "maximum"], label);
-		const minimum = nullableBigint(source.minimum, `${label} minimum`);
-		const maximum = nullableBigint(source.maximum, `${label} maximum`);
-		if (
-			minimum !== null &&
-			maximum !== null &&
-			BigInt(minimum) > BigInt(maximum)
-		)
-			fail(`${label} bounds are invalid`);
-		return Object.freeze({
-			kind: "bigint",
-			minimum,
-			maximum,
-		});
-	}
-	if (source.kind === "numeric") {
-		exact(source, ["kind", "precision", "scale"], label);
-		if (
-			!Number.isSafeInteger(source.precision) ||
-			!Number.isSafeInteger(source.scale) ||
-			(source.precision as number) <= 0 ||
-			(source.precision as number) > 1_000 ||
-			(source.scale as number) < 0 ||
-			(source.scale as number) > (source.precision as number)
-		)
-			fail(`${label} bounds are invalid`);
-		return Object.freeze({
-			kind: "numeric",
-			precision: source.precision as number,
-			scale: source.scale as number,
-		});
-	}
-	if (source.kind === "timestamp") {
-		exact(source, ["kind", "withTimezone"], label);
-		if (typeof source.withTimezone !== "boolean") fail(`${label} is invalid`);
-		return Object.freeze({
-			kind: "timestamp",
-			withTimezone: source.withTimezone,
-		});
-	}
-	fail(`${label} kind is invalid`);
 }
 
 export function postgresTypeForScalarCodec(codec: ScalarCodecV1): string {
